Use term primary/replica instead, deprecate in BC way.
